feat(store user): now can't cancel account with balance left

+++ service/src/main/java/tech/whaleeye/service/Impl/StoreUserServiceImpl.java
@@ -125,6 +125,9 @@ public class StoreUserServiceImpl implements StoreUserService {
 
     @Override
     public Boolean deleteStoreUser(Integer userId) {
+        if (storeUserMapper.getUserById(userId).getAccountBalance().doubleValue() > 0) {
+            throw new InvalidValueException();
+        }
         return storeUserMapper.deleteStoreUser(userId) > 0;
     }
